Move useDisplayErrors to own file
